Temporarily disable active storage
Heroku is giving a couple of warnings relating to active storage in our
build logs as we haven't set this up yet.

```
WARNING:
You set your `config.active_storage.service` to :local in production.
...

WARNING:
We detected that some binary dependencies required to
use all the preview features of Active Storage are not
present on this system.
```

Disabling active storage for the time being, until we provision our
storage provider, so we can keep the build logs quite.