Adds draft key stage page
Ive added in the radio buttons form as per the prototype but I haven't
wired anything up as we're still undecided on how the cpps and keystages
are associated. Also I doubt this page will stay as radio buttons but
instead will likely change to be a list of links as that makes more
sense.