Bypass a GOV.UK tabs bug caused by turbolinks
It appears that when linking to a page with GOV.UK tabs[0] with
turbolinks the contents of the tabs are rendered inline and not
correctly hidden. This is probably due to the way the tabs functionality
is initalised when the page loads.

Disabling turbolinks fixes the problem here but in the longer term (ie
when other pages link directly to lessons too) something more robust
might be in order.

[0] https://design-system.service.gov.uk/components/tabs/