Attempt to simplify default activities with fk
A `lesson_part` has mulitple activities, one of which can be set to
`default` by the supplier. This is now internally represented by a
foreign key on the lesson part, `default_activity_id`, rather than
boolean flags on the activities themselves.

This removes the need for validation on the activities that check their
siblings to ensure they're the only default.