Remove Turbolinks due to GOVUK Tabs bug
Turbolinks appears to cause a bug when used in conjunction with the
GOV.UK Design System tabs. When both are used, the contents of all tabs
are rendered inline on the page.

It would appear that this is caused by the initialisation of tabs not
being triggered by `turbolinks:load` 😔