Disable bullet for activestorage querys
There seems to be an issue with activestorage causing n + 1 querys when
attaching a file. We're already using the provided eagerloading scope
that active storage generates, the issue looks like it's caused by the
way the attach method works.
See this issue on the bullet repo for more details
`https://github.com/flyerhzm/bullet/issues/474`