Adds download controller
Posting to the download controller is intended to kick off the
background job that builds the lesson bundle. The show action renders a
different template depending on the state of the job, we can tringger
polling from the pending template and have a link to download the zip
file in the completed template.

Adds a stub implementation of the job.