Hide download button if nothing to download
In the case that a teacher has chosen activities that don't have any
resources (activities for PE?) then the lesson_bundle zip will be empty.
To avoid this confussion we'll hide the download button if there is
nothing to download and instead only offer the option to print the
lesson plan.