Adds resource models
We want to add previews for open office documents to the app. In order
to avoid requiring libreoffice be installed on the server we intend to
generate the previews at upload time and attach them along with the file
to the resource model.

This commit adds the new resource models and associations to activity.
Note that the associations are prefixed with `temp_` as we don't want to
remove the exisiting activie storage associations just yet. We will once
everything is wired up.