Update Lesson attributes
Lessons data types have been simplified to text fields. This will allow
suppliers more flexibility as they'll be able to format their values as
they wish, in textual format or as a list.

Additionally, summary has been changed to learning_objective, which is
what it's called in the UI. The lengths of name and learning_objective
have been shortened to ensure they display nicely in the UI without
having to overflow or chop.